+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Released on ?.
 - Fixed usage of document ready state (#87) @Sebbs128
 - Fixed `HasChildNodes` is now exposed as a method to DOM (#106) @arekdygas
 - Fixed missing `parent` on Window objects (#68)
+- Fixed missing `NodeFilter` (#101)
 - Fixed XMLHttpRequest handling for relative URLs and failed requests
 - Updated to use AngleSharp v1
 - Updated for Jint v4 (#89, #97) @tomvanenckevort @lahma